requirements.txt
SpeechRecognition
google-generativeai
python-dotenv
streamlit
streamlit-ace
streamlit-extras
pdf2image
gtts
pillow
google.generativeai
IPython
streamlit-lottie
requests
youtube-transcript-api
streamlit-option-menu
streamlit-webrtc
pandas
plotly
Pillow
GitPython
beautifulsoup4
matplotlib
seaborn
spacy
streamlit_shadcn_ui
langchain
PyPDF2
faiss-cpu
poppler-utils
requests_html
streamlit-on-Hover-tabs
httpx-oauth
firebase-admin
streamlit-authenticator
calplot
langchain_google_genai
langchain-community
numpy
aiortc
flask
flask-socketio
streamlit_chat
mysql-connector-python
firebase-admin